Refactor into multiple mods and add modifer tests
This makes the separation of concerns more clear, and should also mean there
does not need to be an explosion of permutations for all teh opcode tests.
